Massive update of all scripts with the new python inifiles
