+++ examples/contamination/so_crosstalk_app.py
@@ -20,14 +20,20 @@ from s4cmb.scanning_strategy import ScanningStrategy
 from s4cmb.tod import TimeOrderedDataPairDiff
 from s4cmb.tod import OutputSkyMap
 
-from s4cmb.config_s4cmb import NormaliseParser
+from s4cmb.config_s4cmb import import_string_as_module
+
+from s4cmb.systematics import inject_crosstalk_inside_SQUID
+
+from s4cmb.xpure import create_batch
+from s4cmb.xpure import write_maps_a_la_xpure
+from s4cmb.xpure import write_weights_a_la_xpure
+import commands
 
 ## Other packages needed
 import os
 import healpy as hp
 import numpy as np
 import argparse
-import ConfigParser
 
 def addargs(parser):
     """ Parse command line arguments for s4cmb """
@@ -63,15 +69,29 @@ def addargs(parser):
         default=None, type=float,
         help='The offset of the HWP in degree.')
 
-    ## Arguments for gain variation - see s4cmb.systematics.
+    ## Arguments for crosstalk - see s4cmb.systematics.
     parser.add_argument(
-        '-gain_variation', dest='gain_variation',
-        default=0.01, type=float,
-        help='Relative gain variation.')
+        '-radius', dest='radius',
+        default=1, type=int,
+        help='Controls the number of bolometers talking within a SQUID.')
+    parser.add_argument(
+        '-mu', dest='mu',
+        default=-0.3, type=float,
+        help='Mean of the Gaussian used to generate \
+        the level of leakage, in percent.')
+    parser.add_argument(
+        '-sigma', dest='sigma',
+        default=0.1, type=float,
+        help='Width of the Gaussian used to generate \
+        the level of leakage, in percent.')
+    parser.add_argument(
+        '-beta', dest='beta',
+        default=2, type=int,
+        help='Exponent controling the attenuation.')
     parser.add_argument(
         '-seed', dest='seed',
         default=5438765, type=int,
-        help='Control the random seed used to generate gain coefficients.')
+        help='Control the random seed used to generate leakage coefficients.')
 
 
 if __name__ == "__main__":
@@ -83,9 +103,8 @@ if __name__ == "__main__":
     addargs(parser)
     args = parser.parse_args(None)
 
-    Config = ConfigParser.ConfigParser()
-    Config.read(args.inifile)
-    params = NormaliseParser(Config._sections['s4cmb'])
+    ## Import parameters from the user parameter file
+    params = import_string_as_module(args.inifile)
 
     ## Overwrite ini file params with params pass to the App directly
     for key in args.__dict__.keys():
@@ -153,13 +172,13 @@ if __name__ == "__main__":
         print("Proc [{}] doing scans".format(rank), range(
             rank, scan.nces, size))
 
-    state = np.random.RandomState(args.seed)
-    new_gains = state.normal(1, args.gain_variation,
-                             size=2 * inst.focal_plane.npair)
+    ## Get SQUID and bolo ID
+    squid_ids = inst.focal_plane.get_indices('Sq')
+    bolo_ids = inst.focal_plane.bolo_index_in_squid
 
     ## Noise seeds
     state_for_noise = np.random.RandomState(params.array_noise_seed)
-    seeds_for_noise = state_for_noise.randint(0, 1e6, scan.nces)
+    seeds_for_noise = np.random.randint(0, 1e6, scan.nces)
     for pos_CES, CESnumber in enumerate(range(rank, scan.nces, size)):
         if params.verbose:
             print("Proc [{}] with seeds ".format(rank),
@@ -172,11 +191,7 @@ if __name__ == "__main__":
             pixel_size=params.pixel_size,
             width=params.width,
             array_noise_level=params.array_noise_level,
-            array_noise_seed=seeds_for_noise[CESnumber],
-            mapping_perpair=True)
-
-        ## Set new gains
-        tod.set_detector_gains(new_gains=new_gains)
+            array_noise_seed=seeds_for_noise[CESnumber])
 
         ## Initialise map containers for each processor
         if pos_CES == 0:
@@ -187,12 +202,21 @@ if __name__ == "__main__":
                                        pixel_size=tod.pixel_size)
 
         ## Scan input map to get TODs
-        for pair in tod.pair_list:
-            d = np.array([
-                tod.map2tod(det) for det in pair])
+        d = np.array([
+            tod.map2tod(det) for det in range(inst.focal_plane.nbolometer)])
 
-            ## Project TOD to maps
-            tod.tod2map(d, sky_out_tot)
+        ## Inject crosstalk
+        inject_crosstalk_inside_SQUID(d,
+                                      squid_ids,
+                                      bolo_ids,
+                                      radius=args.radius,
+                                      mu=args.mu,
+                                      sigma=args.sigma,
+                                      beta=args.beta,
+                                      seed=args.seed)
+
+        ## Project TOD to maps
+        tod.tod2map(d, sky_out_tot)
 
     MPI.COMM_WORLD.barrier()
 
@@ -205,8 +229,6 @@ if __name__ == "__main__":
     sky_out_tot.coadd_MPI(sky_out_tot, MPI=MPI)
 
     if rank == 0:
-        from s4cmb.xpure import write_maps_a_la_xpure
-        from s4cmb.xpure import write_weights_a_la_xpure
         ## Save data on disk into fits file for later use in xpure
         name_out = '{}_{}_{}'.format(params.tag,
                                      params.name_instrument,
@@ -219,11 +241,9 @@ if __name__ == "__main__":
 
         ## Write the submission file for xpure and launch the soft on-the-fly.
         if args.inifile_xpure is not None:
-            from s4cmb.xpure import create_batch
-            import commands
-            Config = ConfigParser.ConfigParser()
-            Config.read(args.inifile_xpure)
-            params_xpure = NormaliseParser(Config._sections['xpure'])
+            ## Import parameters from the user parameter file
+            params_xpure = import_string_as_module(args.inifile_xpure)
+
             batch_file = '{}_{}_{}.batch'.format(
                 params.tag,
                 params.name_instrument,
