+++ examples/contamination/so_flat_MC_diffpointing_app.py
@@ -20,14 +20,15 @@ from s4cmb.scanning_strategy import ScanningStrategy
 from s4cmb.tod import TimeOrderedDataPairDiff
 from s4cmb.tod import OutputSkyMap
 
-from s4cmb.config_s4cmb import NormaliseParser
+from s4cmb.config_s4cmb import import_string_as_module
+
+from s4cmb.systematics import modify_beam_offsets
 
 ## Other packages needed
 import os
 import healpy as hp
 import numpy as np
 import argparse
-import ConfigParser
 
 def safe_mkdir(path):
     """
@@ -53,22 +54,10 @@ def addargs(parser):
     parser.add_argument(
         '-tag', dest='tag',
         required=True,
-        help='Tag name to identify your run. E.g. run_0_crosstalk.')
+        help='Tag name to identify your run. E.g. run_0_nosystematic.')
 
     ## You can also pass any new arguments, or even overwrite those
     ## from the ini file.
-    parser.add_argument(
-        '-type_hwp', dest='type_hwp',
-        default=None,
-        help='The type of HWP that you want to mount on your instrument.')
-    parser.add_argument(
-        '-freq_hwp', dest='freq_hwp',
-        default=None, type=float,
-        help='The frequency of rotation of the HWP in Hz.')
-    parser.add_argument(
-        '-angle_hwp', dest='angle_hwp',
-        default=None, type=float,
-        help='The offset of the HWP in degree.')
     parser.add_argument(
         '-input_filename', dest='input_filename',
         required=True, nargs='+',
@@ -94,15 +83,18 @@ def addargs(parser):
         required=True, type=float,
         help='Name of the output folder.')
 
-    ## Arguments for gain variation - see s4cmb.systematics.
     parser.add_argument(
-        '-gain_variation', dest='gain_variation',
-        default=0.01, type=float,
-        help='Relative gain variation.')
+        '-mu_diffpointing', dest='mu_diffpointing',
+        required=True, type=float,
+        help='Magnitude of the differential pointing (arcsecond).')
     parser.add_argument(
-        '-seed', dest='seed',
-        default=5438765, type=int,
-        help='Control the random seed used to generate gain coefficients.')
+        '-sigma_diffpointing', dest='sigma_diffpointing',
+        required=True, type=float,
+        help='Width of the differential pointing (arcsecond).')
+    parser.add_argument(
+        '-seed_diffpointing', dest='seed_diffpointing',
+        required=True, type=int,
+        help='Seed to generate differential pointing.')
 
 
 if __name__ == "__main__":
@@ -114,9 +106,8 @@ if __name__ == "__main__":
     addargs(parser)
     args = parser.parse_args(None)
 
-    Config = ConfigParser.ConfigParser()
-    Config.read(args.inifile)
-    params = NormaliseParser(Config._sections['s4cmb'])
+    ## Import parameters from the user parameter file
+    params = import_string_as_module(args.inifile)
 
     ## Overwrite ini file params with params pass to the App directly
     for key in args.__dict__.keys():
@@ -136,6 +127,10 @@ if __name__ == "__main__":
 
     if rank == 0:
         safe_mkdir(params.folder_out)
+
+    print('sim [{}]'.format(params.sim_number),
+          params.input_filename,
+          params.array_noise_seed)
 
     ##################################################################
     ## START OF THE SIMULATION
@@ -181,17 +176,30 @@ if __name__ == "__main__":
                             language=params.language)
     scan.run()
 
+    ## Let's inject differential pointing between
+    ## two pixel-pair bolometers in our data!
+    ## The model is the following:
+    ## * Draw from a normal distribution G(mu, sigma)
+    ##      the magnitudes of the differential pointing rho.
+    ## * Draw from a uniform distribution U(0, 2pi) the directions
+    ##      of the differential pointing theta.
+    ## * Move the position of bottom bolometers as
+    ##      - x_top/bottom = \pm rho / 2 * cos(theta)
+    ##      - y_top/bottom = \pm rho / 2 * sin(theta)
+    ## et voila!
+    inst.beam_model.xpos, inst.beam_model.ypos = \
+        modify_beam_offsets(inst.beam_model.xpos,
+                            inst.beam_model.ypos,
+                            mu_diffpointing=args.mu_diffpointing,
+                            sigma_diffpointing=args.sigma_diffpointing,
+                            seed=args.seed_diffpointing)
+
     ## Let's now generate our TOD from our input sky, instrument,
     ## and scanning strategy.
     if params.verbose:
         print("Proc [{}] doing scans".format(rank), range(
             rank, scan.nces, size))
 
-    state = np.random.RandomState(args.seed)
-    new_gains = state.normal(1, args.gain_variation,
-                             size=2 * inst.focal_plane.npair)
-
-    ## Noise seeds
     state_for_noise = np.random.RandomState(params.array_noise_seed)
     seeds_for_noise = state_for_noise.randint(0, 1e6, scan.nces)
     for pos_CES, CESnumber in enumerate(range(rank, scan.nces, size)):
@@ -209,9 +217,6 @@ if __name__ == "__main__":
             array_noise_seed=seeds_for_noise[CESnumber],
             mapping_perpair=True)
 
-        ## Set new gains
-        tod.set_detector_gains(new_gains=new_gains)
-
         ## Initialise map containers for each processor
         if pos_CES == 0:
             sky_out_tot = OutputSkyMap(projection=tod.projection,
@@ -220,12 +225,12 @@ if __name__ == "__main__":
                                        npixsky=tod.npixsky,
                                        pixel_size=tod.pixel_size)
 
-        ## Scan input map to get TODs
+        ## Scan input map to get TODs with original beam offsets
         for pair in tod.pair_list:
             d = np.array([
                 tod.map2tod(det) for det in pair])
 
-            ## Project TOD to maps
+            ## Project TOD to maps with modified beam offsets
             tod.tod2map(d, sky_out_tot)
 
     MPI.COMM_WORLD.barrier()
