+++ examples/perfect/so_flat_app.py
@@ -20,14 +20,13 @@ from s4cmb.scanning_strategy import ScanningStrategy
 from s4cmb.tod import TimeOrderedDataPairDiff
 from s4cmb.tod import OutputSkyMap
 
-from s4cmb.config_s4cmb import NormaliseParser
+from s4cmb.config_s4cmb import import_string_as_module
 
 ## Other packages needed
 import os
 import healpy as hp
 import numpy as np
 import argparse
-import ConfigParser
 
 def addargs(parser):
     """ Parse command line arguments for s4cmb """
@@ -40,66 +39,10 @@ def addargs(parser):
     parser.add_argument(
         '-tag', dest='tag',
         required=True,
-        help='Tag name to identify your run. E.g. run_0_crosstalk.')
-
-    ## Only for xpure use - you do not have to care.
-    parser.add_argument(
-        '-inifile_xpure', dest='inifile_xpure',
-        default=None,
-        help='Configuration file with xpure parameter values.')
+        help='Tag name to identify your run. E.g. run_0_nosystematic.')
 
     ## You can also pass any new arguments, or even overwrite those
     ## from the ini file.
-    parser.add_argument(
-        '-type_hwp', dest='type_hwp',
-        default=None,
-        help='The type of HWP that you want to mount on your instrument.')
-    parser.add_argument(
-        '-freq_hwp', dest='freq_hwp',
-        default=None, type=float,
-        help='The frequency of rotation of the HWP in Hz.')
-    parser.add_argument(
-        '-angle_hwp', dest='angle_hwp',
-        default=None, type=float,
-        help='The offset of the HWP in degree.')
-    parser.add_argument(
-        '-input_filename', dest='input_filename',
-        required=True, nargs='+',
-        help='Input fits with alms.')
-    parser.add_argument(
-        '-array_noise_seed', dest='array_noise_seed',
-        required=True, type=int,
-        help='Seed to generate noise.')
-    parser.add_argument(
-        '-sim_number', dest='sim_number',
-        required=True, type=int,
-        help='Number of the sim.')
-    parser.add_argument(
-        '-folder_out', dest='folder_out',
-        required=True,
-        help='Name of the output folder.')
-    parser.add_argument(
-        '-nside_in', dest='nside_in',
-        required=True, type=int,
-        help='Name of the output folder.')
-    parser.add_argument(
-        '-fwhm_in', dest='fwhm_in',
-        required=True, type=float,
-        help='Name of the output folder.')
-
-    ## Arguments for gain variation - see s4cmb.systematics.
-    parser.add_argument(
-        '-gain_mean', dest='gain_mean',
-        default=1.01, type=float,
-        help='Relative gain variation (mean).')
-    parser.add_argument(
-        '-gain_variation', dest='gain_variation',
-        default=0.01, type=float,
-        help='Relative gain variation (sigma).')
-    parser.add_argument(
-        '-seed', dest='seed',
-        default=5438765, type=int,
-        help='Control the random seed used to generate gain coefficients.')
 
 
 if __name__ == "__main__":
@@ -111,9 +54,8 @@ if __name__ == "__main__":
     addargs(parser)
     args = parser.parse_args(None)
 
-    Config = ConfigParser.ConfigParser()
-    Config.read(args.inifile)
-    params = NormaliseParser(Config._sections['s4cmb'])
+    ## Import parameters from the user parameter file
+    params = import_string_as_module(args.inifile)
 
     ## Overwrite ini file params with params pass to the App directly
     for key in args.__dict__.keys():
@@ -181,11 +123,6 @@ if __name__ == "__main__":
         print("Proc [{}] doing scans".format(rank), range(
             rank, scan.nces, size))
 
-    state = np.random.RandomState(args.seed)
-    new_gains = state.normal(args.gain_mean, args.gain_variation,
-                             size=2 * inst.focal_plane.npair)
-
-    ## Noise seeds
     state_for_noise = np.random.RandomState(params.array_noise_seed)
     seeds_for_noise = state_for_noise.randint(0, 1e6, scan.nces)
     for pos_CES, CESnumber in enumerate(range(rank, scan.nces, size)):
@@ -200,11 +137,7 @@ if __name__ == "__main__":
             pixel_size=params.pixel_size,
             width=params.width,
             array_noise_level=params.array_noise_level,
-            array_noise_seed=seeds_for_noise[CESnumber],
-            mapping_perpair=True)
-
-        ## Set new gains
-        tod.set_detector_gains(new_gains=new_gains)
+            array_noise_seed=seeds_for_noise[CESnumber])
 
         ## Initialise map containers for each processor
         if pos_CES == 0:
@@ -215,12 +148,11 @@ if __name__ == "__main__":
                                        pixel_size=tod.pixel_size)
 
         ## Scan input map to get TODs
-        for pair in tod.pair_list:
-            d = np.array([
-                tod.map2tod(det) for det in pair])
+        d = np.array([
+            tod.map2tod(det) for det in range(inst.focal_plane.nbolometer)])
 
-            ## Project TOD to maps
-            tod.tod2map(d, sky_out_tot)
+        ## Project TOD to maps
+        tod.tod2map(d, sky_out_tot)
 
     MPI.COMM_WORLD.barrier()
 
@@ -233,46 +165,11 @@ if __name__ == "__main__":
     sky_out_tot.coadd_MPI(sky_out_tot, MPI=MPI)
 
     if rank == 0:
-        if params.projection == 'flat':
-            name_out = '{}_{}_{}'.format(params.tag,
-                                         params.name_instrument,
-                                         params.name_strategy)
-            sky_out_tot.pickle_me(
-                '{}/sim{:03d}_{}.pkl'.format(
-                    args.folder_out, args.sim_number, name_out),
-                shrink_maps=False, crop_maps=2**12,
-                epsilon=0., verbose=False)
-
-        elif params.projection == 'healpix':
-            from s4cmb.xpure import write_maps_a_la_xpure
-            from s4cmb.xpure import write_weights_a_la_xpure
-            ## Save data on disk into fits file for later use in xpure
-            name_out = 'sim{:03d}_{}_{}_{}'.format(
-                args.sim_number,
-                params.tag,
-                params.name_instrument,
-                params.name_strategy)
-
-            write_maps_a_la_xpure(sky_out_tot, name_out=name_out,
-                                  output_path='xpure/maps')
-            write_weights_a_la_xpure(sky_out_tot, name_out=name_out,
-                                     output_path='xpure/masks',
-                                     epsilon=0.08, HWP=False)
-
-            if args.inifile_xpure is not None:
-                from s4cmb.xpure import create_batch
-                import commands
-                Config = ConfigParser.ConfigParser()
-                Config.read(args.inifile_xpure)
-                params_xpure = NormaliseParser(Config._sections['xpure'])
-                batch_file = 'sim{:03d}_{}_{}_{}.batch'.format(
-                    args.sim_number,
-                    params.tag,
-                    params.name_instrument,
-                    params.name_strategy)
-                create_batch(batch_file, name_out, params, params_xpure)
-
-                qsub = commands.getoutput('sbatch ' + batch_file)
-                print(qsub)
+        name_out = '{}_{}_{}'.format(params.tag,
+                                     params.name_instrument,
+                                     params.name_strategy)
+        sky_out_tot.pickle_me('sim000_{}.pkl'.format(name_out),
+                              shrink_maps=False, crop_maps=2**12,
+                              epsilon=0., verbose=False)
 
     MPI.COMM_WORLD.barrier()
